vrom/ai_service/requirements.txt
grpcio
grpcio-tools
protobuf
transformers
torch
fastapi
uvicorn
langchain
sentence-transformers
faiss-cpu
langchain-community
